Refactor to remove sklearner (#1361)
